refactor static per tensor and dynamic per token to share base class
